Rename the ApiClientMock into ApiClientMockServer and ApiClientMock
